add some more tests for http app
