+++ http/src/main/scala/weco/http/ErrorDirectives.scala
@@ -2,14 +2,14 @@ package weco.http
 
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server.{Directives, Route}
-import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport
+import de.heikoseeberger.akkahttpcirce.{ErrorAccumulatingCirceSupport}
 import grizzled.slf4j.Logging
 import weco.http.json.DisplayJsonUtil
 import weco.http.models.{ContextResponse, DisplayError}
 
 trait ErrorDirectives
     extends Directives
-    with FailFastCirceSupport
+    with ErrorAccumulatingCirceSupport
     with Logging
     with DisplayJsonUtil {
   import weco.http.models.ContextResponse._
