+++ http/src/test/scala/weco/http/fixtures/HttpFixtures.scala
@@ -1,28 +1,33 @@
 package weco.http.fixtures
 
+import java.net.URL
+
 import org.scalatest.Assertion
 import org.scalatest.matchers.should.Matchers
 import uk.ac.wellcome.monitoring.memory.MemoryMetrics
-import weco.http.monitoring.HttpMetricResults
-
+import weco.http.monitoring.{HttpMetricResults, HttpMetrics}
 import akka.http.scaladsl.Http
 import akka.http.scaladsl.model.HttpMethods.{GET, POST}
-import akka.http.scaladsl.model.{
-  ContentTypes,
-  HttpEntity,
-  HttpRequest,
-  HttpResponse,
-  RequestEntity
-}
+import akka.http.scaladsl.model.{ContentTypes, HttpEntity, HttpRequest, HttpResponse, RequestEntity, StatusCode, StatusCodes}
+import akka.http.scaladsl.server.Route
 import akka.stream.scaladsl.Sink
 import io.circe.parser._
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.matchers.should.Matchers
 import uk.ac.wellcome.akka.fixtures.Akka
 import uk.ac.wellcome.fixtures.TestWith
+import uk.ac.wellcome.json.JsonUtil.toJson
+import uk.ac.wellcome.json.utils.JsonAssertions
+import weco.http.WellcomeHttpApp
+import weco.http.fixtures.ExampleApp.context
 import weco.http.models.HTTPServerConfig
 
-trait HttpFixtures extends Akka with ScalaFutures with Matchers {
+import scala.concurrent.ExecutionContext
+
+trait HttpFixtures extends Akka with ScalaFutures with Matchers
+  with JsonAssertions{
+
+  implicit val ec: ExecutionContext
 
   private def whenRequestReady[R](
                                    r: HttpRequest
@@ -98,4 +103,54 @@ trait HttpFixtures extends Akka with ScalaFutures with Matchers {
     metrics.incrementedCounts should contain(
       s"${name}_HttpResponse_$result"
     )
+
+
+
+  def assertIsDisplayError(
+                            response: HttpResponse,
+                            description: String,
+                            statusCode: StatusCode = StatusCodes.BadRequest
+                          ): Assertion = {
+    response.status shouldBe statusCode
+    response.entity.contentType shouldBe ContentTypes.`application/json`
+
+    withStringEntity(response.entity) { jsonResponse =>
+      assertJsonStringsAreEqual(
+        jsonResponse,
+        s"""
+           |{
+           |  "@context": "$context",
+           |  "errorType": "http",
+           |  "httpStatus": ${statusCode.intValue()},
+           |  "label": "${statusCode.reason()}",
+           |  "description": ${toJson(description).get},
+           |  "type": "Error"
+           |}
+           |""".stripMargin
+      )
+    }
+  }
+
+  def withApp[R](routes: Route)(testWith: TestWith[WellcomeHttpApp, R]): R =
+    withActorSystem { implicit actorSystem =>
+      val metricsName = "example.app"
+
+      val httpMetrics = new HttpMetrics(
+        name = metricsName,
+        metrics = new MemoryMetrics
+      )
+
+      val app = new WellcomeHttpApp(
+        routes = routes,
+        httpMetrics = httpMetrics,
+        httpServerConfig = httpServerConfigTest,
+        contextURL = new URL(context),
+        appName = metricsName
+      )
+
+      app.run()
+
+      testWith(app)
+
+    }
 }
